infosys/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opendeepresearcher"
version = "0.1.0"
description = "Agentic LLM research framework bootstrap"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "langgraph>=0.2.45",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
